feat: Refactor Transpilers to the minimum implementations
